Use maxUsedControlPlaneKubernetesVersion to provide Automatic kube version
