+++ candi/openapi/cluster_configuration.yaml
@@ -122,7 +122,7 @@ apiVersions:
 
           Changing a parameter in a running cluster will [automatically update](https://deckhouse.io/documentation/v1/modules/040-control-plane-manager/#version-control) the cluster's control plane version.
 
-          If `Automatic` is specified, then the control plane version is used, which is considered stable at the moment.
+          If `Automatic` is specified, then the control plane version is used, which is considered stable at the moment. If the stable version of control plane is less than the maximum version that has ever been installed in the cluster, more than 1 minor version, then the version of the cluster will not be changed.
           The version may change when the minor version of the Deckhouse release is changed (see a corresponding release message).
         enum:
         - "1.25"
