+++ candi/openapi/doc-ru-cluster_configuration.yaml
@@ -63,7 +63,8 @@ apiVersions:
 
           Изменение параметра в работающем кластере приведет к [автоматическому обновлению](https://deckhouse.ru/documentation/v1/modules/040-control-plane-manager/#управление-версиями) версии control plane кластера.
 
-          Если указано `Automatic`, то используется версия control plane, которая считается стабильной на данный момент. Изменение этой версии может произойти при смене минорной версии релиза Deckhouse (см. список изменений к релизу).
+          Если указано `Automatic`, то используется версия control plane, которая считается стабильной на данный момент. Если стабильная версия control plane меньше максимальной, которая когда либо была установлена в кластере, больше чем на 1 минорную версию - тогда версия кластера не изменится.
+          Изменение этой версии может произойти при смене минорной версии релиза Deckhouse (см. список изменений к релизу).
       proxy:
         description: |
           Глобальная настройка proxy-сервера (в основном для работы в закрытых окружениях).
